requirements.txt
colorama
discord_py_interactions
numpy
pillow
pydub
python_dateutil
requests
scipy
torch
youtube-dl-nightly